tool: Use a decorator to pass the radio instance
Instead of extracting the radio instance from the Click context, use a
decorator to pass it as a function argument. This means that we can
check for the presence of a radio in the wrapper instead of at
initialization time. A noticeable improvement is that this makes it
possible to access usage strings when the radio isn't connected.

Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>